Separate PainlessLookup into PainlessLookup and PainlessLookupBuilder.
